requirements.txt
flask
torch
transformers
sentence-transformers
unsloth
faiss-cpu
pymupdf
peft
bitsandbytes
